# Improve wpf collection converters performance