… Source/HelixToolkit.Wpf/ExtensionMethods/ConverterExtensions.cs