Hello,
I've modified the extension methods in the Wpf project which convert the float to double and double to float collections. I've improved the performance using the hardware intrinsics (sse and avx). This reduces the execution time for .Net 6 and .Net 8 while still using the same memory size.

Here are the results of a few benchmarks:
```
BenchmarkDotNet v0.14.0, Windows 11 (10.0.22631.4317/23H2/2023Update/SunValley3)
.NET SDK 8.0.403
  Job-LOZNNX : .NET 8.0.10 (8.0.1024.46610), X64 RyuJIT AVX2

InvocationCount=1  IterationCount=1  LaunchCount=1  
UnrollFactor=1  WarmupCount=1  Error=NA  

```
| Method                             | Mean       | Ratio | Allocated | Alloc Ratio |
|----------------------------------- |-----------:|------:|----------:|------------:|
| DoubleToFloat_Current              |  12.153 ms |  1.00 |  38.15 MB |        1.00 |
| DoubleToFloat_New                  |   9.292 ms |  0.76 |  38.15 MB |        1.00 |
|                                    |            |       |           |             |
| FloatToDouble_Current              |  14.040 ms |  1.00 |  76.29 MB |        1.00 |
| FloatToDouble_New                  |  12.788 ms |  0.91 |  76.29 MB |        1.00 |
|                                    |            |       |           |             |
| MeshBuilder_Current                |   8.732 ms |  1.00 |   8.92 MB |        1.00 |
| MeshBuilder_New                    |   2.017 ms |  0.23 |   8.93 MB |        1.00 |
|                                    |            |       |           |             |
| Collection_ToFloat_Current         |  69.686 ms |  1.00 |  38.15 MB |        1.00 |
| Collection_ToFloat_New             |   9.342 ms |  0.13 |  38.15 MB |        1.00 |
|                                    |            |       |           |             |
| Collection_ToInt32_Current         |   5.583 ms |  1.00 |  38.15 MB |        1.00 |
| Collection_ToInt32_New             |   4.600 ms |  0.82 |  38.15 MB |        1.00 |
|                                    |            |       |           |             |
| Collection_ToVector2Point_Current  |  17.807 ms |  1.00 |  15.26 MB |        1.00 |
| Collection_ToVector2Point_New      |   4.805 ms |  0.27 |  15.26 MB |        1.00 |
|                                    |            |       |           |             |
| Collection_ToVector2Vector_Current |  18.171 ms |  1.00 |  15.26 MB |        1.00 |
| Collection_ToVector2Vector_New     |   4.656 ms |  0.26 |  15.26 MB |        1.00 |
|                                    |            |       |           |             |
| Collection_ToVector3Point_Current  |   9.582 ms |  1.00 |  11.44 MB |        1.00 |
| Collection_ToVector3Point_New      |   3.534 ms |  0.37 |  11.45 MB |        1.00 |
|                                    |            |       |           |             |
| Collection_ToVector3Vector_Current |   9.717 ms |  1.00 |  11.44 MB |        1.00 |
| Collection_ToVector3Vector_New     |   3.561 ms |  0.37 |  11.45 MB |        1.00 |
|                                    |            |       |           |             |
| Collection_ToWndDouble_Current     | 165.270 ms |  1.00 |  76.29 MB |        1.00 |
| Collection_ToWndDouble_New         |  13.273 ms |  0.08 |   76.3 MB |        1.00 |
|                                    |            |       |           |             |
| Collection_ToWndInt32_Current      |   5.066 ms |  1.00 |  38.15 MB |        1.00 |
| Collection_ToWndInt32_New          |   4.821 ms |  0.95 |  38.15 MB |        1.00 |
|                                    |            |       |           |             |
| Collection_ToWndPoint_Current      |  44.642 ms |  1.00 |  30.52 MB |        1.00 |
| Collection_ToWndPoint_New          |   8.324 ms |  0.19 |  30.52 MB |        1.00 |
|                                    |            |       |           |             |
| Collection_ToWndPoint3D_Current    |  28.818 ms |  1.00 |  22.89 MB |        1.00 |
| Collection_ToWndPoint3D_New        |   7.047 ms |  0.24 |  22.89 MB |        1.00 |
|                                    |            |       |           |             |
| Collection_ToWndVector3D_Current   |  29.180 ms |  1.00 |  22.89 MB |        1.00 |
| Collection_ToWndVector3D_New       |   7.151 ms |  0.25 |  22.89 MB |        1.00 |